Move main views to their own package
