+++ app/src/main/AndroidManifest.xml
@@ -67,7 +67,7 @@
         </service>
 
         <activity
-            android:name=".ui.CalendarListActivity"
+            android:name=".ui.views.CalendarListActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -75,8 +75,8 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ui.AddCalendarActivity"
-            android:parentActivityName=".ui.CalendarListActivity"
+            android:name=".ui.views.AddCalendarActivity"
+            android:parentActivityName=".ui.views.CalendarListActivity"
             android:exported="true"
             tools:ignore="UnusedAttribute">
             <intent-filter>
@@ -113,14 +113,14 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ui.EditCalendarActivity"
+            android:name=".ui.views.EditCalendarActivity"
             android:label="@string/activity_edit_calendar"
-            android:parentActivityName=".ui.CalendarListActivity"
+            android:parentActivityName=".ui.views.CalendarListActivity"
             android:windowSoftInputMode="stateAlwaysHidden" />
         <activity
             android:name=".ui.InfoActivity"
             android:label="@string/activity_app_info"
-            android:parentActivityName=".ui.CalendarListActivity" />
+            android:parentActivityName=".ui.views.CalendarListActivity" />
 
     </application>
