+++ app/src/main/java/at/bitfire/icsdroid/ProcessEventsTask.kt
@@ -16,7 +16,7 @@ import at.bitfire.icsdroid.calendar.LocalCalendar
 import at.bitfire.icsdroid.calendar.LocalEvent
 import at.bitfire.icsdroid.db.AppDatabase
 import at.bitfire.icsdroid.db.entity.Subscription
-import at.bitfire.icsdroid.ui.EditCalendarActivity
+import at.bitfire.icsdroid.ui.views.EditCalendarActivity
 import at.bitfire.icsdroid.ui.NotificationUtils
 import net.fortuna.ical4j.model.Property
 import net.fortuna.ical4j.model.PropertyList
