+++ app/src/main/java/at/bitfire/icsdroid/ui/views/EditCalendarActivity.kt
@@ -2,7 +2,7 @@
  * Copyright © All Contributors. See LICENSE and AUTHORS in the root directory for details.
  **************************************************************************************************/
 
-package at.bitfire.icsdroid.ui
+package at.bitfire.icsdroid.ui.views
 
 import android.os.Bundle
 import android.widget.Toast
